+fix p

+++ src/ft_p.c
@@ -20,11 +20,11 @@ int		ft_p(t_format_sp spec, va_list ap)
 {
 	void			*p;
 	int 			len;
-	unsigned int	q;
+	unsigned long	q;
 	char 			*s;
 
 	p = va_arg(ap, void *);
-	q = (unsigned int)p;
+	q = (unsigned long)p;
 	spec.flags = 8;
 	s = print_base_nbr(q, 0, spec);
 	s = handler_flags(s, spec);
